ADDED Basic Resume HTML File & Example Template
This commit adds a new html file that is properly styled to display a
resume document and print it out correctly. It sets margins and all
necessary styling for printing.

In addition, there is a lit-html component based template defined in
template.js which actually initializes the different pieces of the
template.

All of these items will be loaded off of a resume.js file which defines
the json for the resume.